+++ src/SteamPanno/panno/loading/PannoLoaderOnline.cs
@@ -60,7 +60,7 @@ namespace SteamPanno.panno.loading
 						{
 							Id = int.Parse(x.Element("appID").Value),
 							Name = x.Element("name").Value,
-							HoursOnRecord = float.Parse(x.Element("hoursOnRecord")?.Value ?? "0", CultureInfo.InvariantCulture),
+							HoursOnRecord = decimal.Parse(x.Element("hoursOnRecord")?.Value ?? "0", CultureInfo.InvariantCulture),
 							HoursOnRecordPrivate = x.Element("hoursOnRecord") == null,
 						}).ToArray();
 					}
